feat: implement Blocknet GPU mining from scratch (CUDA), plus decimal hashrate display
- New from-scratch CUDA implementation of Argon2id (Blake2b + BlaMka compression), written and verified line-by-line against libargon2 (the reference implementation already used and validated for CPU hashing), rather than porting seine's own hand-tuned kernel — prioritizing correctness first over matching their extensive iterative optimization work in one session.
- Cooperative execution: 8 threads per hash (matching the 8 independent rounds in each phase of Argon2's block-mixing function), one hash per CUDA block. Validated to correctly isolate multiple simultaneous hashes with zero cross-contamination via a dedicated 4-hash self-test (blocknet-gpu-selftest), in addition to the original single-hash CPU/GPU comparison test.
- Real-world validated: live shares accepted against the official bntpool.com pool using our own GPU kernel.
- Performance tuning explored this session: maximizing active block count (one hash per block, rather than packing multiple hashes per block) meaningfully improved throughput once we recognized that our tiny batch size (VRAM-limited to ~7 concurrent 2GiB hashes) is far smaller than a modern GPU's SM count, making block-count/SM-distribution the dominant factor over per-block warp occupancy. Shared memory footprint and __launch_bounds__ register-pressure tuning were also attempted; neither moved the needle further.
- Honest status: current GPU throughput (~1.7-3.5 H/s oscillating on an RTX 5070 Ti) has not yet reached parity with seine's own optimized implementation (~4.9 H/s reference measured earlier this session on the same GPU model) or its lower power draw at full utilization. Reaching or exceeding that bar is explicitly deferred to a dedicated future session with proper GPU profiling tools, rather than continuing to guess at further micro-optimizations without one.
- --gpu now has real effect for --algo argon2id-bnt (previously a no-op, CPU-only).

feat: hashrate values (CPU, GPU, total) now shown with one decimal place instead of as integers - important for Blocknet's realistically low hashrate range (often under 10 H/s), where integer-only display could misleadingly oscillate between 0 and 1 without conveying real trend. Also fixes a related display bug: the previous fixed 10-second rate-recompute window could show a false 0 H/s reading whenever a single (slow) Blocknet batch took longer than that window to complete; rate is now recomputed exactly when the hash counter changes, using the real elapsed time since the last change, benefiting any future slow algorithm the same way.

+++ src/monitor/status_table.h
@@ -7,12 +7,12 @@
 struct GpuRow
 {
     GpuStats stats;
-    uint64_t hashrate = 0;
+    double hashrate = 0.0;
 };
 
 struct DashboardData
 {
-    uint64_t totalHashrate = 0;
+    double totalHashrate = 0.0;
     uint64_t shares = 0;
     uint64_t accepted = 0;
     uint64_t rejected = 0;
@@ -20,7 +20,7 @@ struct DashboardData
     uint64_t height = 0;
 
     int cpuThreads = 0;
-    uint64_t cpuHashrate = 0;
+    double cpuHashrate = 0.0;
 
     std::vector<GpuRow> gpuRows;
     std::string algoName;
